♻️ refactor auth screen
